pyproject.toml
[project]
name = "workhuman"
version = "0.1.0"
description = "LLM-based promotion prediction system using LangGraph workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langgraph>=1.0.3",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "pandas>=2.3.3",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.5.0",
    "tqdm>=4.66.0",
]
